fix: move new c509 from cat-voice
