+++ rust/c509-certificate/.config/nextest.toml
@@ -1,3 +1,4 @@
+# cspell: words scrollability testcase
 [store]
 # The directory under the workspace root at which nextest-related files are
 # written. Profile-specific storage is currently written to dir/<profile-name>.
