+++ rust/c509-certificate/deny.toml
@@ -1,4 +1,7 @@
+# cspell: words msvc, wasip, RUSTSEC, rustls, libssh, reqwest, tinyvec, Leay, webpki
+
 [graph]
+# cargo-deny is really only ever intended to run on the "normal" tier-1 targets
 targets = [
     "x86_64-unknown-linux-gnu",
     "aarch64-unknown-linux-gnu",
@@ -22,11 +25,20 @@ ignore = [
 multiple-versions = "warn"
 wildcards = 'deny'
 deny = [
+    # { crate = "git2", use-instead = "gix" },
     { crate = "openssl", use-instead = "rustls" },
     { crate = "openssl-sys", use-instead = "rustls" },
     "libssh2-sys",
+    # { crate = "cmake", use-instead = "cc" },
+    # { crate = "windows", reason = "bloated and unnecessary", use-instead = "ideally inline bindings, practically, windows-sys" },
 ]
-skip = []
+skip = [
+    # { crate = "bitflags@1.3.2", reason = "https://github.com/seanmonstar/reqwest/pull/2130 should be in the next version" },
+    # { crate = "winnow@0.5.40", reason = "gix 0.59 was yanked, see https://github.com/Byron/gitoxide/issues/1309" },
+    # { crate = "heck@0.4.1", reason = "strum_macros uses this old version" },
+    # { crate = "base64@0.21.7", reason = "gix-transport pulls in this old version, as well as a newer version via reqwest" },
+    # { crate = "byte-array-literalsase64@0.21.7", reason = "gix-transport pulls in this old version, as well as a newer version via reqwest" },
+]
 skip-tree = [
     { crate = "windows-sys@0.48.0", reason = "a foundational crate for many that bumps far too frequently to ever have a shared version" },
 ]
@@ -45,7 +57,9 @@ allow-git = [
 
 [licenses]
 version = 2
-unused-allowed-license = "allow"
+# Don't warn if a listed license isn't found
+unused-allowed-license="allow"
+# We want really high confidence when inferring licenses from text
 confidence-threshold = 0.93
 allow = [
     "MIT",
@@ -60,7 +74,11 @@ allow = [
     "Unicode-3.0",
     "MPL-2.0",
 ]
-exceptions = []
+exceptions = [
+    #{ allow = ["Zlib"], crate = "tinyvec" },
+    #{ allow = ["Unicode-DFS-2016"], crate = "unicode-ident" },
+    #{ allow = ["OpenSSL"], crate = "ring" },
+]
 
 [[licenses.clarify]]
 crate = "byte-array-literals"
@@ -76,3 +94,24 @@ license-files = [{ path = "../LICENSE-MIT", hash = 0x001c7e6c }]
 crate = "ring"
 expression = "MIT"
 license-files = [{ path = "LICENSE", hash = 0xbd0eed23 }]
+
+# SPDX considers OpenSSL to encompass both the OpenSSL and SSLeay licenses
+# https://spdx.org/licenses/OpenSSL.html
+# ISC - Both BoringSSL and ring use this for their new files
+# MIT - "Files in third_party/ have their own licenses, as described therein. The MIT
+# license, for third_party/fiat, which, unlike other third_party directories, is
+# compiled into non-test libraries, is included below."
+# OpenSSL - Obviously
+#expression = "ISC AND MIT AND OpenSSL"
+#license-files = [{ path = "LICENSE", hash = 0xbd0eed23 }]
+
+#[[licenses.clarify]]
+#crate = "webpki"
+#expression = "ISC"
+#license-files = [{ path = "LICENSE", hash = 0x001c7e6c }]
+
+# Actually "ISC-style"
+#[[licenses.clarify]]
+#crate = "rustls-webpki"
+#expression = "ISC"
+#license-files = [{ path = "LICENSE", hash = 0x001c7e6c }]
