+++ rust/c509-certificate/src/general_names/data.rs
