+++ rust/c509-certificate/src/general_names/general_name.rs
@@ -13,7 +13,7 @@ use super::{
     data::{get_gn_from_int, get_gn_value_type_from_int, get_int_from_gn},
     other_name_hw_module::OtherNameHardwareModuleName,
 };
-use crate::{c509_name::Name, c509_oid::C509oid};
+use crate::{name::Name, oid::C509oid};
 
 /// A struct represents a `GeneralName`.
 /// ```cddl
@@ -173,7 +173,8 @@ impl Encode<()> for GeneralNameValue {
     }
 }
 impl<C> Decode<'_, C> for GeneralNameValue
-where C: GeneralNameValueTrait + Debug
+where
+    C: GeneralNameValueTrait + Debug,
 {
     fn decode(d: &mut Decoder<'_>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
         match ctx.get_type() {
@@ -197,11 +198,9 @@ where C: GeneralNameValueTrait + Debug
                 let value = Name::decode(d, &mut ())?;
                 Ok(GeneralNameValue::Name(value))
             },
-            GeneralNameValueType::Unsupported => {
-                Err(minicbor::decode::Error::message(
-                    "Cannot decode Unsupported GeneralName value",
-                ))
-            },
+            GeneralNameValueType::Unsupported => Err(minicbor::decode::Error::message(
+                "Cannot decode Unsupported GeneralName value",
+            )),
         }
     }
 }
@@ -242,9 +241,10 @@ mod test_general_name {
         let mut buffer = Vec::new();
         let mut encoder = Encoder::new(&mut buffer);
 
-        let hw = OtherNameHardwareModuleName::new(oid!(2.16.840 .1 .101 .3 .4 .2 .1), vec![
-            0x01, 0x02, 0x03, 0x04,
-        ]);
+        let hw = OtherNameHardwareModuleName::new(
+            oid!(2.16.840 .1 .101 .3 .4 .2 .1),
+            vec![0x01, 0x02, 0x03, 0x04],
+        );
         let gn = GeneralName::new(
             GeneralNameTypeRegistry::OtherNameHardwareModuleName,
             GeneralNameValue::OtherNameHWModuleName(hw),
